Split benchmark code into cpp and header
